app/requirements.txt
cassandra-driver==3.28.0
venv-pack==0.2.0
pyarrow==11.0.0
fastparquet==2022.12.0
pathvalidate==2.5.2
pyspark==3.5.5
tqdm==4.65.0
